Javadocs, move commands to packages
