+++ src/main/java/net/cytonic/cytosis/commands/CommandHandler.java
@@ -1,9 +1,16 @@
 package net.cytonic.cytosis.commands;
 
 import net.cytonic.cytosis.Cytosis;
+import net.cytonic.cytosis.commands.defaultMinecraft.GamemodeCommand;
+import net.cytonic.cytosis.commands.defaultMinecraft.TeleportCommand;
 import net.cytonic.cytosis.commands.moderation.BanCommand;
+import net.cytonic.cytosis.commands.server.CreateInstanceCommand;
+import net.cytonic.cytosis.commands.server.PodDetailsCommand;
+import net.cytonic.cytosis.commands.server.ShutdownInstancesCommand;
+import net.cytonic.cytosis.commands.server.StopCommand;
 import net.minestom.server.command.CommandManager;
 import net.minestom.server.entity.Player;
+
 import java.util.Scanner;
 import java.util.concurrent.Executors;
 import java.util.concurrent.ScheduledExecutorService;
