+++ src/main/java/net/cytonic/cytosis/commands/server/CreateInstanceCommand.java
@@ -1,4 +1,4 @@
-package net.cytonic.cytosis.commands;
+package net.cytonic.cytosis.commands.server;
 
 import net.cytonic.cytosis.Cytosis;
 import net.cytonic.cytosis.config.CytosisSettings;
@@ -6,8 +6,14 @@ import net.minestom.server.command.builder.Command;
 
 import static net.cytonic.cytosis.utils.MiniMessageTemplate.MM;
 
+/**
+ * A command to create a new instance
+ */
 public class CreateInstanceCommand extends Command {
 
+    /**
+     * A command to create a new instance
+     */
     public CreateInstanceCommand() {
         //todo: specify different kinds of instances
         super("createinstance");
