+++ src/main/java/net/cytonic/cytosis/commands/server/ShutdownInstancesCommand.java
@@ -1,4 +1,4 @@
-package net.cytonic.cytosis.commands;
+package net.cytonic.cytosis.commands.server;
 
 import net.cytonic.cytosis.Cytosis;
 import net.cytonic.cytosis.config.CytosisSettings;
@@ -6,7 +6,14 @@ import net.minestom.server.command.builder.Command;
 
 import static net.cytonic.cytosis.utils.MiniMessageTemplate.MM;
 
+/**
+ * A command for closing all instances of cytosis on a kubernetes cluster
+ */
 public class ShutdownInstancesCommand extends Command {
+
+    /**
+     * A command to close all instances
+     */
     public ShutdownInstancesCommand() {
         super("shutdowninstances");
         setCondition((sender, _) -> sender.hasPermission("cytosis.commands.shutdowninstances"));
